Add BeaconsResponseFactory to build the DTO for response
